+++ src/App.jsx
@@ -2,11 +2,9 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'; 
 import LandingPage from './Pages/Landing/Landing'
 import AboutPage from './Pages/About/About'
-import NavMenu from './Pages/NavMenu/NavMenu';
 import CoffeeMenuPage from './Pages/CoffeeMenu/CoffeeMenu'
 import ProfilePage from './Pages/Profile/Profile'
 import StatusPage from './Pages/Status/Status'
-import Cart from './Pages/Cart/Cart'
 
 function App() {
   console.log("Rendering App component");
@@ -16,11 +14,9 @@ function App() {
       <Routes>
         <Route path="/" element={<LandingPage />} /> 
         <Route path="/about" element={<AboutPage />} />
-        <Route path="/navmenu" element={<NavMenu />} />
         <Route path="/coffee-menu" element={<CoffeeMenuPage />} /> 
         <Route path="/profile" element={<ProfilePage />} /> 
-        <Route path="/status" element={<StatusPage />} /> 
-        <Route path="/cart" element={<Cart />}/>  
+        <Route path="/status" element={<StatusPage />} />  
       </Routes>
     </Router>
   )
