+++ src/Components/Footer/Footer.css
@@ -1,0 +1,7 @@
+.footer {
+    width: 100%;
+    height: 73px;
+    background-image: url('../../assets/header.svg');
+    background-size: contain;
+    transform: rotate(180deg);
+}
