+++ src/Components/Footer/Footer.jsx
@@ -1,0 +1,8 @@
+import headerImage from "../../assets/header.svg"; 
+import './Footer.css'
+
+export default function Footer() {
+    return (
+        <footer className="footer" />
+    )
+}
