+++ src/Components/Header/Header.jsx
@@ -1,27 +1,47 @@
-import navIcon from "../../assets/navicon.svg";
-import { Link } from "react-router-dom";
-import cartIcon from "../../assets/cart.svg";
+import { useState } from 'react'
+
+import reactangle from "../../assets/menu_icon.svg";
+
 import "./header.css";
 import useStore from "../../store/store";
+import CartButton from "../CartButton/CartButton"
+import NavMenu from "../NavMenu/NavMenu"
+import Cart from "../Cart/Cart"
 
 const Header = () => {
+  const [navIsOpen, setNavIsOpen] = useState(false)
+
   const { cart } = useStore();
   //för att räkna antal produkter som finns i cart
   const totalQuantity = cart.reduce(
     (total, cartItem) => total + cartItem.quantity,
     0
   );
-  return (
-    <div className="headerContainer">
-      <Link className="navIcon" to="/nav">
-        <img src={navIcon} alt="Navigation Icon" />
-      </Link>
 
-      <Link className="cartIcon" to="/cart">
-        <img className="cartLogo" src={cartIcon} alt="Cart Icon" />
-        <p className="quantityHolder">{totalQuantity}</p>
-      </Link>
-    </div>
+
+  const toggleNavigation = () => {
+    if (navIsOpen) {
+      setNavIsOpen(false)
+    } else {
+      setNavIsOpen(true)
+    }
+  }
+
+  if (navIsOpen) {
+    return (
+      <NavMenu onClose={toggleNavigation} />
+    )
+  }
+
+  return (
+    <>
+      <header className="headerContainer">
+        <button className="navButton" onClick={toggleNavigation} >
+          <img src={reactangle} className="navIcon" alt="Navigation Icon" />
+        </button>
+
+      </header>
+    </>
   );
 };
