+++ src/Pages/About/About.jsx
@@ -1,18 +1,19 @@
 import "./about.css";
-import aboutImage from"../../assets/about.png"
-import MenuButton from "../../Components/MenuButton/MenuButton"
+
+import Header from "../../Components/Header/Header";
+import Footer from "../../Components/Footer/Footer";
 
 function About() {
-  const style = {
-    backgroundImage: `url(${aboutImage})`,
-    backgroundSize: "cover",
-    height: "100vh",
-  };
 
   return (
-    <div className="about" style={style}>
-<MenuButton />
-    </div>
+    <>
+      <Header />
+      <div className="about" >
+        
+      </div>
+      <Footer />
+    </>
+
   )
 }
