+++ src/Pages/CoffeeMenu/CoffeeMenu.css
@@ -1,47 +1,14 @@
 .menu {
-    background-color: #f3e4e1;
     display: flex;
     flex-direction: column;
     justify-content: space-between;
-    height: 100vh; /* Gör så att sidan tar upp hela höjden */
-}
-
-.menuIcon {
-    position: absolute;
-    top: 17px;
-    left: 39px;
-    width: 55px;
-    height: 55px;
-}
-
-.cartIcon {
-    width: 25px;
-    height: 25px;
-}
-.cartButton {
-    position: absolute;
-    top: 30px;
-    right: 32px;
-    padding: 2rem;
-    z-index: 1000;
-    background-color: black;
-    border-radius: 50%;
-    width: 40px;
-    height: 40px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.menuHeader,
-.menuFooter {
-    width: 100%;
-    text-align: center;
-    padding: 10px;
 }
 
 /* Justera header så att den inte är förskjuten nedåt */
 .menuHeader {
+    width: 100%;
+    text-align: center;
+    padding: 10px;
     font-family: PT Serif;
     margin-top: 0; /* Ta bort eller sätt till 0 för att inte förskjuta headern */
     font-size: 2.6rem;
@@ -90,9 +57,4 @@ ul {
     font-size: 0.8rem;
     margin-top: -1rem;
     margin-left: 0.5rem;
-}
-
-/* Footern ska vara uppochnedvänd */
-.footerImage {
-    transform: rotate(180deg);
 }
