+++ src/index.css
@@ -6,11 +6,15 @@
 body {
     font-size: 16px;
     color: #333;
+    margin: 0 auto;
+    width: 100%;
+    max-width: 370px;
+    overflow-x: hidden;
+    display: flex;
 }
 
 #root {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
+    width: 100%;
+    background-color: #f3e4e1;
+    position: relative;
 }
